pyproject.toml
[tool.poetry]
name = "ecommerce_backend"
version = "0.1.0"
description = ""
authors = ["Leonam Silva <leonam.silva.cc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "2.2.15"
psycopg2-binary = "^2.8.5"
djangorestframework = "^3.11.1"
autopep8 = "^1.5.4"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.10.1"
pylint = "^2.6.0"
flake8 = "^3.8.3"
isort = "^5.4.2"
pytest-django = "^3.9.0"
factory_boy = "^3.0.1"
pytest-randomly = "^3.4.1"
ipdb = "^0.13.3"
pdbpp = "^0.10.2"
freezegun = "^0.3.15"
black = "^20.8b1"
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
